+++ .eslintrc.js
@@ -89,11 +89,9 @@ module.exports = {
       files: [ "sitesupport/vue/*.js",
                "console/vue.js",
                "modules/concom/report/sitesupport/vue.js",
-               "modules/concom/sitesupport/vue.js",
                "modules/concom/sitesupport/division-parser.js",
-               "modules/concom/sitesupport/components/*.js",
-               "modules/concom/sitesupport/concom-list.js",
                "modules/concom/sitesupport/department-staff-parser.js",
+               "modules/concom/vue/**/*.js",
                "modules/event/report/sitesupport/vue.js",
                "modules/registration/sitesupport/badgeMenuPane.js" ,
                "modules/registration/sitesupport/vue.js",
